Added longpolling recovery logic:
The longpolling cycle can be broken if a response from the server is
missed ( usually due to user computer being off/asleep ) and this
results in the client code to never be able to recover and reconnect to
the server itself ). The longpolling timeout is added to detect this
issue and reconnect automatically.